Allow using 32b integer as the tensor dimension/index type
As a side effect, this patch also makese the dimension width
type of an explicit fixed size. The previously used size_t is
device-dependent and a typical 64b width is clearly overkill for
most of the use cases. An option to use 32b data type adds more
data level parallelism for computations with tensor dimension
data as well as simplifies address computations.

The patch adds a new typedef glow::dim_t with a signed counterpart
glow::sdim_t which are defined to a fixed width 32b type when
cmake -DTENSOR_DIMS_32_BITS=ON.