Add more inference timing stats to the ImageClassifier
When benchmarking long-running inference scenarions it is (sometimes)
interesting to:

a) Skip the first run from the statistics. This is interesting in case
there is aggressive specialization or other preperation in the first run
which won't recur in the later runs. This is the case at least with
OpenCL when applying kernel specialization which incurs several kernel
online builds for the first run, but the rest of the runs get the built
programs from an in-memory cache.

b) The best runtime. Especially when benchmarking on a desktop computer
with other background tasks that interfere with the execution, the best
runtime of N runs gives a better indication of the achievable peak
performance than the average performance.

This patch adds the ability to exclude the first warmup run from the
average run computation and adds a separate statistics that records
the best run out of all warmup runs.